Title: 2개씩 그룹짓기 2, Time: 58ms, Memory: 1MB, Status: Runtime Error - CodeTree

+++ 250204/2개씩 그룹짓기 2/group-of-pairs-2.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main() {
     // Please write your code here.
-    int arr[100000];
+    long long arr[100000];
     
     int n;
     cin >> n;
@@ -13,7 +13,7 @@ int main() {
     
     sort(arr, arr+2*n);
 
-    int min_val = 1 << 30;
+    long long min_val = 1 << 30;
     for (int i=0; i<n; i++) {
         min_val = min(min_val, arr[i+n] - arr[i]);
     }
@@ -21,22 +21,3 @@ int main() {
     
     return 0;
 }
-
-// 2
-
-// 2n
-// 15   2   7   9   10   5
-
-// 2   5   7   9   10   15
-
-// 2, 5 = 3
-
-
-// 2, 7 = 5
-// 2, 9
-// 2, 10 = 8
-// 5, 7, 9, 15
-
-
-// 2, 15
-
